+++ bindings/nostr-js/src/nips/nip44.rs
@@ -12,16 +12,12 @@ use crate::key::{JsPublicKey, JsSecretKey};
 
 #[wasm_bindgen(js_name = NIP44Version)]
 pub enum JsNIP44Version {
-    /// V1 - Deprecated
-    Deprecated = 1,
     V2 = 2,
 }
 
 impl From<Version> for JsNIP44Version {
     fn from(version: Version) -> Self {
         match version {
-            #[allow(deprecated)]
-            Version::V1 => Self::Deprecated,
             Version::V2 => Self::V2,
         }
     }
@@ -30,8 +26,6 @@ impl From<Version> for JsNIP44Version {
 impl From<JsNIP44Version> for Version {
     fn from(version: JsNIP44Version) -> Self {
         match version {
-            #[allow(deprecated)]
-            JsNIP44Version::Deprecated => Self::V1,
             JsNIP44Version::V2 => Self::V2,
         }
     }