+++ CHANGELOG.md
@@ -68,6 +68,7 @@
 
 * Remove deprecated methods/functions ([Yuki Kishimoto])
 * nostr: remove support for `nrelay` NIP-19 entity ([Yuki Kishimoto])
+* nostr: remove support for NIP-44 v1 ([Yuki Kishimoto])
 * cli: remove `tracing-subscriber` dep
 
 ## [v0.33.0]